requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
plotly>=5.0.0
sqlalchemy>=1.4.0
pymysql>=1.1.0
python-dotenv>=0.19.0
requests>=2.28.0